Fix integration issues

+++ eth-scan/src/main/kotlin/co/nilin/opex/chainscan/eth/impl/EthereumInterpreter.kt
@@ -5,6 +5,7 @@ import co.nilin.opex.chainscan.core.spi.Interpreter
 import org.springframework.stereotype.Component
 import org.web3j.protocol.core.methods.response.EthBlock
 import java.math.BigDecimal
+import java.math.BigInteger
 
 const val ERC20_TRANSFER_METHOD_SIG = "0xa9059cbb000000000000000000000000"
 const val ETH_TRANSFER_METHOD_SIG = "0x"
@@ -23,7 +24,7 @@ class EthereumInterpreter : Interpreter<EthBlock.TransactionObject> {
     private fun parseTokenTransfer(tx: EthBlock.TransactionObject): Transfer {
         if (!isTokenTransfer(tx.input)) throw IllegalArgumentException()
         val receiver = "0x${tx.input.substring(34, 74)}"
-        val amount = BigDecimal(tx.input.substring(74).toLong(16).toBigInteger(), 18)
+        val amount = BigDecimal(BigInteger(tx.input.substring(74), 16), 18)
         return Transfer(tx.hash, tx.from, receiver, true, tx.to, amount)
     }
